app/requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
bounded-pool-executor==0.0.3
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
easyocr==1.5.0
fastjsonschema==2.15.3
idna==3.3
imageio==2.19.3
imutils==0.5.4
jsonschema==4.6.1
jupyter-core==4.10.0
nbformat==5.4.0
networkx==2.8.4
numpy==1.23.0
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.2.0
pybind11==2.9.2
pycountry==22.3.5
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.9
python-bidi==0.4.2
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
soupsieve==2.3.2.post1
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.9
./bin/fasttext-0.9.2-cp310-cp310-win_amd64.whl